Move proxy code to an example
It is an example, after all, and it feels like it belongs in a separate
space.
