+++ Cargo.toml
@@ -20,3 +20,6 @@ thiserror = "1.0.38"
 arbitrary = { version = "1.3.2", features = ["derive"] }
 arbtest = "0.3.1"
 expect-test = "1.5.0"
+
+[[example]]
+name = "proxy"
